+++ tests/test.ts
@@ -8,6 +8,9 @@ const expected2 = Deno.readTextFileSync("tests/expected2.html");
 Deno.test("Render TSX", async () => {
   const code = await example();
   assertEquals(await renderComponent(code), expected);
+  assertEquals(await renderComponent(code), await code.toString());
+
   const code2 = await example2();
   assertEquals(await renderComponent(code2), expected2);
+  assertEquals(await renderComponent(code2), await code2.toString());
 });